requirements.txt
Flask==2.2.2
scikit-surprise==1.1.3
numpy
pandas
kafka==1.3.5               
kafka-python==2.0.2
requests
